backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
google-genai
django-cors-headers
python-dotenv
djangorestframework-simplejwt
